+++ CHANGES.md
@@ -8,7 +8,7 @@ Changelog
 
 - Configuration field `keep_timeboxes` to configure the maximum number
   of timeboxes to retain in state per user per channel.
-- Configuration field `keep_duration` to configure the maximum
+- Configuration field `keep_duration_seconds` to configure the maximum
   duration for which timeboxes are retained in state.
 - Configuration field `max_print_channel` to configure the maximum
   number of timeboxes to be listed in a channel in response to `list`
@@ -16,6 +16,15 @@ Changelog
 - Configuration field `max_print_private` to configure the maximum
   number of timeboxes to be listed in private message in response to
   `list` or `mine` commands.
+- Configuration field `default_duration_minutes` to configure the
+  default duration of a timebox.
+- Configuration field `duration_multiple_minutes` to configure the
+  number that the duration of a timebox (in minutes) must be a
+  multiple of.
+- Configuration field `min_duration_minutes` to configure the minimum
+  allowed duration of a timebox.
+- Configuration field `max_duration_minutes` to configure the maximum
+  allowed duration of a timebox.
 
 
 0.2.0 (2024-08-03)